+++ models/models.go
@@ -168,7 +168,7 @@ type (
 	// See Phone Numbers, Formatting for more information. If needed, On-Premises API users can get this number by
 	// calling the contacts' endpoint.
 	//
-	// Type type (string). Optional. The type of message you want to send. Default: text.
+	// Type (string). Optional. The type of message you want to send. Default: text.
 	Message struct {
 		Product       string       `json:"messaging_product"`
 		To            string       `json:"to"`
