+++ qrcodes/qr.go
@@ -93,7 +93,7 @@ func Create(ctx context.Context, client *http.Client, rtx *RequestContext,
 	return &response, nil
 }
 
-func List(ctx context.Context, client *gphttp.Client, rctx *RequestContext, hooks ...whttp.Hook) (*ListResponse, error) {
+func List(ctx context.Context, client *http.Client, rctx *RequestContext, hooks ...whttp.Hook) (*ListResponse, error) {
 	reqCtx := &whttp.RequestContext{
 		Name:       "list qr codes",
 		BaseURL:    rctx.BaseURL,
